+++ fixtures/uitests/tests/ui/enum_all_variants_skipped.rs
@@ -1,0 +1,9 @@
+fn main() {} /* empty main required by `trybuild` */
+
+#[derive(uniffi::Enum)]
+pub enum AllSkipped {
+    #[uniffi(skip)]
+    Hidden(std::sync::mpsc::Receiver<()>),
+}
+
+uniffi_macros::setup_scaffolding!();