+++ CHANGELOG.md
@@ -18,6 +18,9 @@
   See [#2911](https://github.com/mozilla/uniffi-rs/pull/2911).
 - Added support for remote trait interfaces - ie, traits defined in a crate which doesn't use
   UniFFI. Use `#[uniffi::export(remote)]` or `[Trait, Remote]` in UDL. Foreign implementations are not supported, see the docs for more.
+- Added `#[uniffi(skip)]` for enum and error variants. The variant is left out
+  of the generated bindings entirely, so its fields can hold Rust types that
+  can't cross the FFI.
 
 ### What's Fixed
 - Kotlin: Fixed messages for error classes that inherit `Throwable`, but not `Exception`.